[stdlib] Pattern matching: introduce EncodedCollection
Instead of being able to match collections directly, we now match
EncodedCollections, which carry an encoding and a collection of native
code units.

We need a way for Strings, which will be collections of graphemes, to
represent the fact that they have an encoding so that pattern matchers
can take advantage of that information.